Search.html, style.css updates
Updates for the search and style files. I have made and attend to modify the search html with a image. I have not updated the text styling yet. please provide your and input that you fee could be valuable.  I made modifications to the style sheet, and added an image to search.html.